consolidate some of the source files
